+++ tests/profiles/dsig_test.py
@@ -2,7 +2,7 @@ from fontbakery.codetesting import (assert_PASS,
                                     assert_results_contain,
                                     CheckTester,
                                     TEST_FILE)
-from fontbakery.checkrunner import FAIL
+from fontbakery.checkrunner import WARN
 from fontbakery.profiles import opentype as opentype_profile
 
 from fontTools.ttLib import TTFont
@@ -13,15 +13,13 @@ def test_check_dsig():
     check = CheckTester(opentype_profile,
                         "com.google.fonts/check/dsig")
 
-    # Our reference Cabin Regular font is good (theres a DSIG table declared):
+    # Our reference Cabin Regular font is bad (theres a DSIG table declared):
     ttFont = TTFont(TEST_FILE("cabin/Cabin-Regular.ttf"))
+    assert_results_contain(check(ttFont),
+                           WARN, 'found-DSIG',
+                           'with a font containing a DSIG table...')
 
-    # So it must PASS the check:
+    # Then we remove the DSIG table and it should now PASS the check:
+    del ttFont['DSIG']
     assert_PASS(check(ttFont),
                 'with a good font...')
-
-    # Then we remove the DSIG table so that we get a FAIL:
-    del ttFont['DSIG']
-    assert_results_contain(check(ttFont),
-                           FAIL, 'lacks-signature',
-                           'with a font lacking a DSIG table...')